QCamera: Interface cleanup
- All size member variables changed to size_t.
  Others changed to named size types, e.g
  int changed to int32_t, etc.
- Moved macro CAM_DUMP_TO_FILE to
  'mm_camera_interface.h'.
- Removed irrelevant macro FLIP_V_H

CRs-Fixed: 641033

+++ QCamera2/HAL3/QCamera3Channel.cpp
@@ -300,11 +300,11 @@ int32_t QCamera3Channel::stop()
     ATRACE_CALL();
     int32_t rc = NO_ERROR;
     if(!m_bIsActive) {
-        ALOGE("%s: Attempt to stop inactive channel",__func__);
+        ALOGE("%s: Attempt to stop inactive channel", __func__);
         return rc;
     }
 
-    for (int i = 0; i < m_numStreams; i++) {
+    for (uint8_t i = 0; i < m_numStreams; i++) {
         if (mStreams[i] != NULL) {
             mStreams[i]->stop();
         }
@@ -331,7 +331,7 @@ int32_t QCamera3Channel::stop()
 int32_t QCamera3Channel::bufDone(mm_camera_super_buf_t *recvd_frame)
 {
     int32_t rc = NO_ERROR;
-    for (int i = 0; i < recvd_frame->num_bufs; i++) {
+    for (uint32_t i = 0; i < recvd_frame->num_bufs; i++) {
          if (recvd_frame->bufs[i] != NULL) {
              for (int j = 0; j < m_numStreams; j++) {
                  if (mStreams[j] != NULL &&
@@ -2904,7 +2904,7 @@ int32_t QCamera3ReprocessChannel::extractFrameAndCrop(mm_camera_super_buf_t *fra
         return BAD_VALUE;
     }
 
-    for (int i = 0; i < frame->num_bufs; i++) {
+    for (uint32_t i = 0; i < frame->num_bufs; i++) {
 
         QCamera3Stream *pStream = getStreamBySrcHandle(frame->bufs[i]->stream_id);
         QCamera3Stream *pSrcStream = getSrcStreamBySrcHandle(frame->bufs[i]->stream_id);
